+++ Config.lua
@@ -853,18 +853,21 @@ DF.PartyDefaults = {
 
     -- Direct Mode: Buff Filters
     directBuffShowAll = false,                -- Show all buffs (ignores sub-filters)
-    directBuffFilterPlayer = true,            -- PLAYER filter
+    directBuffFilterPlayer = false,           -- PLAYER filter
     directBuffFilterRaid = false,             -- RAID filter
     directBuffFilterRaidInCombat = true,      -- RAID_IN_COMBAT filter
     directBuffFilterCancelable = false,       -- CANCELABLE filter
+    directBuffFilterNotCancelable = false,    -- NOT_CANCELABLE filter
     directBuffFilterImportant = false,        -- IMPORTANT filter (12.0.1)
+    directBuffFilterBigDefensive = true,      -- BIG_DEFENSIVE filter (12.0.1)
+    directBuffFilterExternalDefensive = true, -- EXTERNAL_DEFENSIVE filter (12.0.0)
     directBuffSortOrder = "DEFAULT",          -- "DEFAULT" / "TIME" / "NAME"
 
     -- Direct Mode: Debuff Filters
-    directDebuffShowAll = true,               -- Show all debuffs (ignores sub-filters)
-    directDebuffFilterRaid = false,           -- RAID filter
-    directDebuffFilterCrowdControl = false,   -- CROWD_CONTROL filter
-    directDebuffFilterImportant = false,      -- IMPORTANT filter (12.0.1)
+    directDebuffShowAll = false,              -- Show all debuffs (ignores sub-filters)
+    directDebuffFilterRaid = true,            -- RAID filter
+    directDebuffFilterCrowdControl = true,    -- CROWD_CONTROL filter
+    directDebuffFilterImportant = true,       -- IMPORTANT filter (12.0.1)
     directDebuffSortOrder = "DEFAULT",        -- "DEFAULT" / "TIME" / "NAME"
 
     buffGrowth = "LEFT_UP",
@@ -2014,18 +2017,21 @@ DF.RaidDefaults = {
 
     -- Direct Mode: Buff Filters
     directBuffShowAll = false,                -- Show all buffs (ignores sub-filters)
-    directBuffFilterPlayer = true,            -- PLAYER filter
+    directBuffFilterPlayer = false,           -- PLAYER filter
     directBuffFilterRaid = false,             -- RAID filter
     directBuffFilterRaidInCombat = true,      -- RAID_IN_COMBAT filter
     directBuffFilterCancelable = false,       -- CANCELABLE filter
+    directBuffFilterNotCancelable = false,    -- NOT_CANCELABLE filter
     directBuffFilterImportant = false,        -- IMPORTANT filter (12.0.1)
+    directBuffFilterBigDefensive = true,      -- BIG_DEFENSIVE filter (12.0.1)
+    directBuffFilterExternalDefensive = true, -- EXTERNAL_DEFENSIVE filter (12.0.0)
     directBuffSortOrder = "DEFAULT",          -- "DEFAULT" / "TIME" / "NAME"
 
     -- Direct Mode: Debuff Filters
-    directDebuffShowAll = true,               -- Show all debuffs (ignores sub-filters)
-    directDebuffFilterRaid = false,           -- RAID filter
-    directDebuffFilterCrowdControl = false,   -- CROWD_CONTROL filter
-    directDebuffFilterImportant = false,      -- IMPORTANT filter (12.0.1)
+    directDebuffShowAll = false,              -- Show all debuffs (ignores sub-filters)
+    directDebuffFilterRaid = true,            -- RAID filter
+    directDebuffFilterCrowdControl = true,    -- CROWD_CONTROL filter
+    directDebuffFilterImportant = true,       -- IMPORTANT filter (12.0.1)
     directDebuffSortOrder = "DEFAULT",        -- "DEFAULT" / "TIME" / "NAME"
 
     buffGrowth = "LEFT_UP",